renamed appropriate test functions
